% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_team_Record.R
\name{all_team_Record}
\alias{all_team_Record}
\title{Extract all teams' win/loss records and winning percentages for the games recorded in tibble}
\usage{
all_team_Record(data)
}
\arguments{
\item{data}{This is the tibble of data from kenpom.com}
}
\value{
This function returns a tibble summarizing wins, losses, and a
win percentage for all teams in a tibble
}
\description{
This will take in a tibble and return a tibble with the win/loss record of
all teams in the tibble.  This is a nested function requiring team_record.R.
}
\examples{
## filter game data 
a tibble with four columns
data <- tibble(
 team1 = rep(c("A", "B", "C", "D"), each = 4), 
 team2 = rep(c("A", "B", "C", "D"), times = 4),
 Margin = runif(16, min = -75, max = 75))
 
 all_team_Record(data)
 

}
