% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byTeamGames.R
\name{byTeamGames}
\alias{byTeamGames}
\title{Identify all games where a given team is a player (whether home team or visiting)}
\usage{
byTeamGames(data, team, ungroup)
}
\arguments{
\item{data}{This is the tibble of data from kenpom.com}

\item{team}{The string name of the team, i.e.:  team = 'Indiana State'}

\item{ungroup}{If TRUE will undo some other previous grouping}
}
\value{
This function returns tibble containing rows of all home or visiting
games for a given team.
}
\description{
This will take in a tibble and the string name of a given team and return a
tibble with all games played by that team.
}
\examples{
## filter game data 
a tibble with four columns
data <- tibble(
 team1 = rep(c("A", "B", "C", "D"), each = 4), 
 score1 = runif(16, min = 0, max = 100),        
 team2 = rep(c("A", "B", "C", "D"), times = 4),
 score2 = runif(16, min = 0, max = 100))
 
 byTeamGames(data, 'A')
 

}
