% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginHist.R
\name{marginHist}
\alias{marginHist}
\title{Make a histogram of victory margins (win/loss)}
\usage{
marginHist(data, divideBinsBy = 1)
}
\arguments{
\item{data}{This is the tibble of data from kenpom.com}

\item{divideBinsBy}{This allows user to reduce default number of bins,
default argument is no division (1).}
}
\value{
This function returns a tibble summarizing wins, losses, and a
win percentage for all teams in a tibble
}
\description{
This will take in a tibble and return a tibble a colored histogram of the
victory margins of games played.
}
\examples{
## make a plotted histgram: 
a tibble with four columns
data <- tibble(
 team1 = rep(c("A", "B", "C", "D"), each = 4), 
 team2 = rep(c("A", "B", "C", "D"), times = 4),
 Margin = runif(16, min = -75, max = 75))
 
 plot <- marginHist(data)
 plot
 
 

}
