% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginMonthlyRange.R
\name{marginMonthlyRange}
\alias{marginMonthlyRange}
\title{Make a histogram of victory margins (win/loss)}
\usage{
marginMonthlyRange(data, months = c(11, 12, 1, 2))
}
\arguments{
\item{data}{This is the tibble of data from kenpom.com}

\item{months}{This allows user to specify the numeric int months in the order
they wish them to appear in the box plot:  a vector argument, default is
c(11,12,1,2), the months of the basketball season.}
}
\value{
This function returns a boxplot showing the ranges of margins over the
months of the basketball season.
}
\description{
This will take in a tibble and return a tibble a colored histogram of the
victory margins of games played.
}
\examples{
## make a boxplot
a tibble with four columns
data <- tibble(
 team1 = rep(c("A", "B", "C", "D"), each = 4), 
 team2 = rep(c("A", "B", "C", "D"), times = 4),
 Margin = runif(16, min = -75, max = 75))
 
 plot <- marginMonthlyRange(data)
 plot
 
 

}
