% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCollBBTeam.R
\name{runcollBBTeam}
\alias{runcollBBTeam}
\title{College BB Team App}
\usage{
runcollBBTeam()
}
\description{
This function allows the college team shiny app to run.  This app offers users
a team-by-team summary of statistics, based on their entry, presented in a
table with a supplementary plot.
}
