% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_Record.R
\name{team_Record}
\alias{team_Record}
\title{Extract a team's win/loss record and winning percentage for the games recorded in tibble}
\usage{
team_Record(data, team)
}
\arguments{
\item{data}{This is the tibble of data from kenpom.com}

\item{team}{The string name of the team, i.e.:  team = 'Indiana State'}
}
\value{
This function returns a tibble summarizing wins, losses, and a win percentage
}
\description{
This will take in a tibble and the string name of a given team and return a
tibble with the win/loss record of that given team
}
\examples{
## filter game data 
a tibble with four columns
data <- tibble(
 team1 = rep(c("A", "B", "C", "D"), each = 4), 
 team2 = rep(c("A", "B", "C", "D"), times = 4),
 Margin = runif(16, min = -75, max = 75))
 
 team_Record(data, 'A')
 

}
